style: fix padding and margin
