+++ src/routes/index.ts
@@ -2,4 +2,4 @@ export * from "./welcome";
 export * from "./onboarding";
 export * from "./registration";
 export * from "./login";
-export * from "./settings"
+export * from "./settings";
